Android: Start workers if pending jobs exist, refactor startWorker logic.
Modified worker initiation process in FileTransferBackground.java to ensure workers start if there are pending jobs at initialization. Previously, workers were only if an upload was added. Any jobs that were in the queue were ignored until then.

Additionally, we can now just call startWorkers() and it will check to see if the workers are already started. If they are, do nothing.